Further TimescaleDB improvements (#285)
- Add config settings for Beta Read from/Write to TimescaleDB (will be removed eventually)
- Add high service ranking for Influx (will be removed eventually)
- Improve Exception handling in TimescaleDB write
- Fix parseLong bug in JsonUtils
- Improve `fems-release.sh` script
- Fix version information in `user.component.html`
- Drop unused `doc/modules/ROOT/pages/single_document.adoc`
- Drop obsolete `tools/prepare-next-snapshot-patch.sh`

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/285

+++ io.openems.common/src/io/openems/common/utils/JsonUtils.java
@@ -851,7 +851,7 @@ public class JsonUtils {
 		}
 		if (jPrimitive.isString()) {
 			var string = jPrimitive.getAsString();
-			return Integer.parseInt(string);
+			return Long.parseLong(string);
 		}
 		throw OpenemsError.JSON_NO_NUMBER.exception(jPrimitive.toString().replace("%", "%%"));
 	}
